Defining the Model
The model category we are dealing with is a classification model (binary classification), so the model is defined in the following way. Here, two tokens [CLS] and [SEP] are also defined. CLS is used to output the features that express the entire sentence, and SEP is used to segment the sentence. When doing multi-classification, we can choose to use the output vector of CLS for fine-tuning, or use the vector of the entire sequence for further processing.