Defining datasets
The problem dealt with here is to determine the similarity between two sentences. Very similar to tf.dataset, the dataset here is responsible for outputting the input data in the format required by the model. 
Data preprocessing can be added here, and its function is similar to an advanced generator.