Defining training process
Since it is not convenient to output AUC when calling the trainer API (actually because the documentation is too complicated and I can't find ready-made code), I referred to the code shared by others on the Internet and completed it. Using the tqdm interface, 

Note that the input of the model is ids, attention mask, and label. Since I have never used BERT before, I did not notice the importance of the attention mask input (BERT only calculates the loss of the masked part). I adjusted it many times before the code ran through...